tf/vsphere: move ip extract code to vsphere stage package
Move the code for finding the IP address of a vsphere VM
from pkg/terraform/gather/vsphere to pkg/terraform/stages/vsphere
so that it is located in the same place where it is being used.
